Added features to UserHeader/Edit Profile
Added custom badge, custom website/URL, and custom instagram spots for users to add to their headers/profile. This included adding a few files in server/utils/helpers that worked as scripts for adding these fields to the MongoDB database. Added redirects for the Cancel and Submit buttons in Edit User Profile page as well for convenience.