Added Teabag Tab!
This took a lot of console logging to finalize... posts were not showing up correctly for different suer types (own posts, tagged posts, and teabags). This was because the data structure returned from the server was different from what the client expected. It was also due to using a single state variable for all types of posts which made it difficult to separate them. AND the fetching logic initially did not account for new post types which was an oversight. So general posts state was replaced with separate state variables for userPosts, taggedPosts, and teabagPosts. Then the getPosts function was modified to handle the 3 types of posts. Then the rending logic was updated so each tab uses its corresponding state logic. And finally, adjusted useEffect hook to fetch all post types when the componenet mounts or username changes.