Merged HomediggitiesREFACTORv2, ready to deploy!
Merged HomediggitiesREFACTORv2 which is the Welcome Screen and Trivia Component to get into the application. The idea for this is to eventually have a main account that is part of different circles/spheres that has customizable features such as a Welcome screen, Trivia Component, and other circle/sphere based customizables. This is an example of a working circle/sphere, although with the final product the authentication and login order will be different and more fluid.