Optimized TriviaComponent, new triviaData, changed Teabag to Frame.
The incorrectimage and incorrectsound on triviacomponent / page was a little slow on mobile browser, so I implemented a way to preload the images when a user is on the page. Added some actual trivia questions to demo what kind of questions one might customize for their own circle. Changed Teabag to Frame since a friend had a good point about that word meaning something different in the gaming community.